Enhance the Link component
- Rename it to LinkAnt
- Introduce the optional `widgetStyle` prop.
  If you pass in "button", it will render and AntD button instead of a link.
- There is also `buttonProps`, which are props that are passed down directly to the
  button. (For styling.)
