+++ examples/src/menus/MoreMenusAnt.tsx
@@ -1,7 +1,7 @@
 import {
     MenuAndContentAnt,
     TextFormAnt,
-    Link,
+    LinkAnt,
     ArgChangingLink,
     MultiArgChangingLink,
     UIFragmentSpec,
@@ -38,7 +38,12 @@ export class MoreMenusAnt extends React.Component<{ url?: UrlStore } & Navigable
                         <div>And here is a text field</div>
                         <TextFormAnt operation={url!.bindSearch} />
                         <hr />
-                        <Link to={['moreMenus', 'two']} label="Select 'two' on the left tab menu!" />
+                        <LinkAnt
+                            widgetStyle="button"
+                            to={['moreMenus', 'two']}
+                            label="Select 'two' on the left tab menu!"
+                            buttonProps={{ type: 'primary' }}
+                        />
                         <hr />
                         <ArgChangingLink arg={url!.color!} value="blue" label="Set the color to blue!" />
                         <hr />
