+++ packages/antd/src/not-antd/index.ts
@@ -1,4 +1,3 @@
-export { Link } from './Link';
 export { ArgChangingLink } from './ArgChangingLink';
 export { MultiArgChangingLink } from './MultiArgChangingLink';
 export { LocationDependentArea, LocationDependentAreaProps } from './LocationDependentArea';
