+++ packages/antd/src/routing/index.ts
@@ -1,3 +1,4 @@
+export { LinkAnt } from './LinkAnt';
 export { NavMenuBarAnt } from './NavMenuBarAnt';
 export { MenuAndContentAnt } from './MenuAndContentAnt';
 export { NavStepsAnt } from './NavStepsAnt';
